fixing type mappings
