+++ test/js/plugin.test.ts
@@ -1,5 +1,5 @@
 // This is a base test for the plugin interface.
-import { MrujsPluginInterface } from '../../src/types'
+import { MrujsPluginInterface } from '../../types'
 import { assert } from '@esm-bundle/chai'
 import { spy } from 'sinon'
