+++ test/js/plugins/cableCar.ts
@@ -3,7 +3,7 @@ import { assert } from '@esm-bundle/chai'
 import CableReady from 'cable_ready'
 
 import mrujs from '../../../src'
-import { CableCar } from '../../../src/plugins'
+import { CableCar } from '../../../plugins/'
 
 const cableCar = new CableCar(CableReady)
