+++ tsconfig-lint.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["./src", "./test/js"]
+  "include": ["./types.ts", "./src", "./plugins", "./test/js"]
 }
