+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["./src"],
+  "include": ["./types.ts", "./src"],
   "exclude": ["./dist"],
   "compilerOptions": {
     "baseUrl": ".",
